[ADD] tgfx 코드 반영.

+++ 03.src/app/TouchGFX/gui/include/gui/compass_screen/CompassPresenter.hpp
@@ -28,7 +28,7 @@ public:
     virtual ~CompassPresenter() {}
 
     // override the empty function.
-    void notifyHDGValue(int hdgValue);
+    void notifyHDGValue(float hdgValue);
 
 private:
     CompassPresenter();
